Moved some things out of jsapi-require (#79-1)
